fix: corrigir tipo INTEGER na RPC fetch_pending_jobs e build frontend no railway.json
- supabase-rpc-queue.sql: id BIGINT → INTEGER e v_selected_ids BIGINT[] → INTEGER[]
  (correspondendo ao tipo real da tabela ghl_wa_message_queue — resolvia erro 42804)
- railway.json: buildCommand inclui build do frontend React/Vite
  (Dockerfile já resolve via multistage, mas alinha documentação)

+++ supabase-rpc-queue.sql
@@ -7,7 +7,7 @@ DROP FUNCTION IF EXISTS fetch_pending_jobs(INT);
 
 CREATE OR REPLACE FUNCTION fetch_pending_jobs(batch_size INT)
 RETURNS TABLE (
-  id BIGINT,
+  id INTEGER,
   instance_id TEXT,
   type TEXT,
   to_number TEXT,
@@ -21,7 +21,7 @@ RETURNS TABLE (
   updated_at TIMESTAMPTZ
 ) AS $$
 DECLARE
-  v_selected_ids BIGINT[];
+  v_selected_ids INTEGER[];
 BEGIN
   -- 1. Selecionar IDs para travar (Locking)
   WITH locked_rows AS (
